fix(regression): instructions and translate will now correctly send drafts.id when running on a published document
This fixes for a regression introduced in 3.1.0.

Note: Updating is strictly not necessary to get the fix, as the backend will now always create a draft using the action api, if the document has a public-id and is not liveEdit: true. However, sending the drafts.id is still more correct, as it show the correct intent of the client. Assist should never mutate published documents.